Removendo class antiga e refatorando código
